refactor: 로그인 컴포넌트들을 메인 페이지로 통합
